+++ Changes.md
@@ -5,6 +5,7 @@
 HEAD
 -----------
 
+- Work around error running Web UI against Redis Cluster [#3492]
 - Remove core extensions, Sidekiq is now monkeypatch-free! [#3474]
 - Reimplement Web UI's HTTP\_ACCEPT\_LANGUAGE parsing because the spec is utterly
   incomprehensible for various edge cases. [johanlunds, natematykiewicz, #3449]